project/mini/atm/atm_flask/requirements.txt
Flask>=2.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
bcrypt>=4.0.0
cryptography>=41.0.0